+++ src/main/java/com/nttdatacasefirst/stockAPI/repository/CapitalIncreaseRepository.java
@@ -4,11 +4,12 @@ import com.nttdatacasefirst.stockAPI.entity.CapitalIncrease;
 import org.springframework.stereotype.Repository;
 
 import java.awt.print.Pageable;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface CapitalIncreaseRepository extends BaseRepository<CapitalIncrease>  {
     CapitalIncrease findByArrangementNo(Long arrNo);
     CapitalIncrease findByYear(int year);
-    List<CapitalIncrease> findByCapitalValueLessThan(int value);
+    List<CapitalIncrease> findByCapitalValueLessThan(BigDecimal value);
 }
